refactor(svelte): break apart a bit
BREAKING CHANGE: Having both `<style type="text/m-css">` **and** a `<link>` element in a component will no longer throw. `<style>` takes precedence over `<link>`.
